Random Row Sample

Validates that the number of columns is equal to expected count

RowSample(x, n = 5)

Arguments

x
a matrix-like R object

n
the number of random rows to display

Value

If object row count exceeds supplied parameter or default value, then the function returns a random sample of "n" rows. If object row count is less than parameter or default, the function returns all rows of the object.

See also

Examples

if (FALSE) {
#Obtain a Random Sample of Rows
}
x = matrix(rnorm(100), ncol = 10)
RowSample(x)

#>            [,1]       [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,] -1.8193472 0.07202570 -0.1692801 -1.3928471  1.73357811  1.8034834
#> [2,]  0.6161543 1.17966418  0.4078885 -1.1353433 -0.08007363 -0.3840073
#> [3,] -0.4004052 0.04213411 -1.4761969  1.2663406  1.70460874  1.5181931
#> [4,]  0.5591060 1.47753231  1.2083856 -0.1109194  0.52645799  0.8130582
#> [5,] -1.1694615 0.34391334 -1.9795551  2.0866886 -0.14542689  1.0628765
#>            [,7]        [,8]         [,9]      [,10]
#> [1,] -0.9754230  0.03476606 -0.704819663  1.0174911
#> [2,] -1.7133026  0.40510118  0.007586777 -1.3416861
#> [3,]  0.9824534  1.15234707  0.020831228  2.0341899
#> [4,] -0.9604492 -0.09855069  1.840316741 -0.9615715
#> [5,] -0.6984294 -0.34708847  0.934261130 -0.8452478